requirements.txt
flask
flask-socketio
flask-login
pymongo
werkzeug
python-dotenv
gunicorn
eventlet
gevent
waitress